feat(support_management): add engineer signature field to maintenance visit
Added a new custom field 'Engineer Signature' to the Maintenance Visit doctype to facilitate capturing the engineer's signature post-visit. Adjusted the field order and indexes to integrate this new field properly. Also, updated the default print format to 'Engineer Visit Custom' to include this new field in the printed document.

Signed-off-by: Grawish Sachdeva <41504211+grawish@users.noreply.github.com>